push

+++ app/Http/Controllers/Admin/PermissionController.php
@@ -20,7 +20,7 @@ class PermissionController extends Controller
      */
     public function index()
     {
-        // abort_if(Gate::denies('permission_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('permission_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $permission = Permission::query()->paginate(5);
         return view('admin.permissions.index', compact('permission'));
@@ -32,7 +32,7 @@ class PermissionController extends Controller
      */
     public function create()
     {
-        // abort_if(Gate::denies('permission_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('permission_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         return view('admin.permissions.create');
     }
@@ -47,8 +47,7 @@ class PermissionController extends Controller
     {
         Permission::create($request->all());
 
-        return redirect()->route('permissions.index');
-        return back()->with('success', 'Permission created successfully');
+        return redirect()->route('permissions.index')->with('success', 'Permission created successfully!');
     }
 
     /**
@@ -59,7 +58,7 @@ class PermissionController extends Controller
      */
     public function show(Permission $permission)
     {
-        // abort_if(Gate::denies('permission_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('permission_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         return view('admin.permissions.show', compact('permission'));
     }
@@ -72,7 +71,7 @@ class PermissionController extends Controller
      */
     public function edit(Permission $permission)
     {
-        // abort_if(Gate::denies('permission_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('permission_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         return view('admin.permissions.edit', compact('permission'));
     }
@@ -88,8 +87,7 @@ class PermissionController extends Controller
     {
         $permission->update($request->all());
 
-        return redirect()->route('permissions.index');
-        return back()->with('success', 'Permission updated successfully');
+        return redirect()->route('permissions.index')->with('success', 'Permission updated successfully');
     }
 
     /**
@@ -100,7 +98,7 @@ class PermissionController extends Controller
      */
     public function destroy(Permission $permission)
     {
-        // abort_if(Gate::denies('permission_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('permission_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $permission->delete();
         return back()->with('success', 'Permission deleted successfully');
